requirements.txt
# Core
aiopg==0.10.0
peewee==2.8.3
peewee-async==0.5.5
psycopg2==2.6.2
tornado==4.4.2
jsonschema==2.5.1
msgpack-python==0.4.8

# Development & Testing
py==1.4.31
pytest==3.0.2
pytest-cov==2.3.1
pytest-env==0.6.0
pytest-tornado==0.4.5
coverage==4.2